Redirect: Only redirect GET. Save desired location
Do not redirect POST. This could cause some undesired behaviour. Just fail.
This means the user should load some authenticated page before calling APIs.

Save the page we want to visit in #, and on the JavaScript side, save it
in sessionStorage. This should allow multiple tabs to be redirected BACK to
the page they were going to.