Fail sign in page over HTTP unless explicitly permitted
This makes it easy to debug why localhost sign ins are not working. We mark
the cookie as "secure", so if you are using http://localhost it won't work.
This makes the request fail obviously with a useful message on the server.

Rename InsecureCookies -> PermitInsecureCookies() to make it clear it is
configuring the Authenticator.